scrapers/requirements.txt
selenium
webdriver-manager
supabase
python-dotenv
requests
